% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dkextra}
\alias{dkextra}
\title{Extra Desikan-Killiany Cortical Atlas}
\usage{
dkextra()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Extra Desikan-Killiany Cortical Atlas
}
\seealso{
Other ggseg_atlases: 
\code{\link{hcpa}()}
}
\concept{ggseg_atlases}
