% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegDefaultExtra-package.R
\docType{package}
\name{ggsegDefaultExtra-package}
\alias{ggsegDefaultExtra}
\alias{ggsegDefaultExtra-package}
\title{ggsegDefaultExtra: Extra dk and aseg datasets for the ggseg-plotting tool}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is a support package for the ggseg, and ggseg3d packages. It contains the extra dk and aseg, with some minor differences from the original atlases to plot using functions from those two packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegDefaultExtra}
  \item Report bugs at \url{https://github.com/ggseg/ggsegDefaultExtra/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
