% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcpa}
\alias{hcpa}
\title{Human Connectome Project Subcortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\usage{
hcpa
}
\description{
Human Connectome Project Subcortical Atlas
}
\keyword{datasets}
