% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dkextra}
\alias{dkextra}
\title{Extra Desikan-Killiany Cortical Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\usage{
dkextra
}
\description{
Extra Desikan-Killiany Cortical Atlas
}
\keyword{datasets}
