% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hcpa}
\alias{hcpa}
\title{Human Connectome Project Subcortical Atlas}
\usage{
hcpa()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Human Connectome Project Subcortical Atlas
}
